Visualize the Posterior Distributions of Model Differences

A density is created for each contrast in a facetted grid.

# S3 method for posterior_diff
ggplot(data, mapping = NULL, ...,
  environment = NULL, size = 0)

Arguments

data
An object produced by contrast_models().

mapping, ..., environment
Not currently used.

size
The size of an effective difference. For example, a 5% increase in accuracy between two models might be considered a "real" difference.

Value

A ggplot2::ggplot() object using geom_density facetted by the models being contrasted (when there are 2 or more contrasts).

Examples

# Example objects from the "Getting Started" vignette at
#  https://topepo.github.io/tidyposterior/articles/Getting_Started.html

# File for pre-run model is at
ex_dat <- "https://bit.ly/2S1v6H9"

# load(load(url(ex_dat))

# library(ggplot2)
# ggplot(glm_v_nnet, size = 0.02) + theme_bw()