Summarize Posterior Distributions of Model Differences

Credible intervals are created for the differences. Also, region of practical equivalence (ROPE) statistics are computed when the effective size of a difference is given.

# S3 method for posterior_diff
summary(object, prob = 0.9, size = 0, ...)

Arguments

object
An object produced by contrast_models().

prob
A number p (0 < p < 1) indicating the desired probability mass to include in the intervals.

size
The size of an effective difference. For example, a 5 "real" difference.

...
Not currently used

Value

A data frame with interval and ROPE statistics for each comparison.

Details

The ROPE estimates included in the results are the columns pract_neg, pract_equiv, and pract_pos. pract_neg integrates the portion of the posterior below -size (and pract_pos is the upper integral starting at size). The interpretation depends on whether the metric being analyzed is better when larger or smaller. pract_equiv integrates between [-size, size]. If this is close to one, the two models are unlikely to be practically different relative to size.