Extract Posterior Distributions for Models

tidy can be used on an object produced by perf_mod() to create a data frame with a column for the model name and the posterior predictive distribution values.

# S3 method for perf_mod
tidy(x, seed = sample.int(10000, 1), ...)

Arguments

x
An object from perf_mod()

seed
A single integer for sampling from the posterior.

...
Not currently used

Value

A data frame with the additional class "posterior"

Details

Note that this posterior only reflects the variability of the groups (i.e. the fixed effects). This helps answer the question of which model is best for this data set. If does not answer the question of which model would be best on a new resample of the data (which would have greater variability).

Examples

# Example from "Getting Started" vignette at
#  https://topepo.github.io/tidyposterior/articles/Getting_Started.html

file <- system.file("examples", "roc_model.RData", package = "tidyposterior")
load(file)

roc_model

#> Bayesian Analysis of Resampling Results
#> Original data: 10-fold cross-validation using stratification
#> 

posterior_values <- tidy(roc_model)
head(posterior)

#> Error in head(posterior): object 'posterior' not found

class(posterior)

#> Error in eval(expr, envir, enclos): object 'posterior' not found