Example Data Sets

Example Data Sets

Value

Tibbles with the additonal class rset

Details

Several data sets are contained in the package as examples. Each simulates an rset object but the splits columns are not included to save space.

precise_example contains the results of the classification analysis of a real data set using 10-fold CV. The holdout data sets contained thousands of examples and have precise performance estimates. Three models were fit to the original data and several performance metrics are included.

noisy_example was also generated from a regression data simulation. The original data set was small (50 samples) and 10-repeated of 10-fold CV were used with four models. There is an excessive of variability in the results (probably more than the resample-to-resample variability). The RMSE distributions show fairly right-skewed distributions.

concrete_example contains the results of the regression case study from the book Applied Predictive Modeling. The original data set contained 745 samples in the training set. 10-repeats of 10-fold CV was also used and 13 models were fit to the data.

ts_example is from a data set where rolling-origin forecast resampling was used. Each assessment set is the summary of 14 observations (i.e. 2 weeks). The analysis set consisted of a base of about 5,500 samples plus the previous assessment sets. Four regression models were applied to these data.

Examples

data(precise_example)
precise_example

#> #  10-fold cross-validation using stratification 
#> # A tibble: 10 x 29
#>    splits     id glm_Accuracy glm_Kappa   glm_ROC  glm_Sens  glm_Spec glm_PRAUC
#>     <lgl>  <chr>        <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1     NA Fold01    0.7219789 0.3276231 0.7980770 0.7294282 0.7202908 0.4890519
#>  2     NA Fold02    0.6964700 0.2896545 0.7779284 0.7196653 0.6912137 0.4559959
#>  3     NA Fold03    0.7012887 0.2972104 0.7900645 0.7234637 0.6962705 0.4862322
#>  4     NA Fold04    0.7041999 0.3157213 0.7947942 0.7625698 0.6909953 0.4974863
#>  5     NA Fold05    0.7210201 0.3235922 0.7972035 0.7224547 0.7206951 0.4807800
#>  6     NA Fold06    0.7108993 0.3032914 0.7804242 0.7057183 0.7120733 0.4835726
#>  7     NA Fold07    0.7023963 0.3048452 0.7899755 0.7391911 0.6940582 0.4850924
#>  8     NA Fold08    0.7175258 0.3211274 0.7838771 0.7290503 0.7149178 0.4765777
#>  9     NA Fold09    0.7196599 0.3280866 0.7953167 0.7391911 0.7152339 0.4910643
#> 10     NA Fold10    0.7194022 0.3235179 0.7960815 0.7280335 0.7174463 0.4876692
#> # ... with 21 more variables: glm_Precision <dbl>, glm_Recall <dbl>,
#> #   glm_F <dbl>, knn_Accuracy <dbl>, knn_Kappa <dbl>, knn_ROC <dbl>,
#> #   knn_Sens <dbl>, knn_Spec <dbl>, knn_PRAUC <dbl>, knn_Precision <dbl>,
#> #   knn_Recall <dbl>, knn_F <dbl>, nnet_Accuracy <dbl>, nnet_Kappa <dbl>,
#> #   nnet_ROC <dbl>, nnet_Sens <dbl>, nnet_Spec <dbl>, nnet_PRAUC <dbl>,
#> #   nnet_Precision <dbl>, nnet_Recall <dbl>, nnet_F <dbl>