Summarize the Posterior Distributions of Model Statistics

Numerical summaries are created for each model including the posterior mean and upper and lower credible intervals (aka uncertainty intervals).

# S3 method for posterior
summary(object, prob = 0.9,
  seed = sample.int(10000, 1), ...)

Arguments

object
An object produced by tidy.perf_mod().

prob
A number p (0 < p < 1) indicating the desired probability mass to include in the intervals.

seed
A single integer for sampling from the posterior.

...
Not currently used

Value

A data frame with summary statistics and a row for each model.

Examples

# Example objects from the "Getting Started" vignette at
#  https://topepo.github.io/tidyposterior/articles/Getting_Started.html

file <- system.file("examples", "roc_model.RData", package = "tidyposterior")
load(file)

posterior_values <- tidy(roc_model)
summary(posterior_values)

#> [38;5;246m# A tibble: 3 x 4[39m
#>   model  mean lower upper
#>   [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m glm   0.790 0.786 0.795
#> [38;5;250m2[39m knn   0.753 0.748 0.757
#> [38;5;250m3[39m nnet  0.841 0.837 0.846