Summarize the Posterior Distributions of Model Statistics

Numerical summaries are created for each model including the posterior mean and upper and lower credible intervals (aka uncertainty intervals).

# S3 method for posterior
summary(object, prob = 0.9, seed = sample.int(10000, 1), ...)

Arguments

object
An object produced by tidy.perf_mod().

prob
A number p (0 < p < 1) indicating the desired probability mass to include in the intervals.

seed
A single integer for sampling from the posterior.

...
Not currently used

Value

A data frame with summary statistics and a row for each model.

Examples

# Example objects from the "Getting Started" vignette at
#  https://topepo.github.io/tidyposterior/articles/Getting_Started.html

# File for pre-run model is at
ex_dat <- "https://bit.ly/2OJdvl1"

# load(load(url(ex_dat))

# posterior_values <- tidy(roc_model)
# summary(posterior_values)